move octicons to deprecated
