+++ packages/styled-react/src/deprecated.tsx
@@ -1,6 +1,6 @@
 export {TabNav, type TabNavProps, type TabNavLinkProps} from './components/deprecated/TabNav'
 export {Dialog, type DialogProps, type DialogHeaderProps} from './components/deprecated/DialogV1'
-export {Octicon, type OcticonProps} from './components/Octicon'
+export {Octicon, type OcticonProps} from './components/deprecated/Octicon'
 export {Tooltip, type TooltipProps} from './components/deprecated/Tooltip'
 export {
   ActionList,
